requirements.txt
sentence_transformers
einops
pyobjc-framework-Contacts
lancedb
pyarrow
pandas